requirements.txt
flask~=3.1.0
kafka-python~=2.0.2
psycopg2-binary~=2.9.10
psycopg2~=2.9.10

# OpenTelemetry SDK and Instrumentations
opentelemetry-distro~=0.52b0
opentelemetry-api~=1.31.0
opentelemetry-sdk~=1.31.0
opentelemetry-instrumentation~=0.52b0
opentelemetry-instrumentation-requests~=0.52b0
opentelemetry-instrumentation-flask~=0.52b0
opentelemetry-instrumentation-psycopg2~=0.52b0
opentelemetry-instrumentation-kafka-python~=0.52b0

# OpenTelemetry Exporter (Matching SDK)
opentelemetry-exporter-otlp~=1.31.0
